+++ src/OpenCOVER/OpenVRUI/base/vruiCollabInterface.cpp
@@ -101,7 +101,7 @@ int vruiCollabInterface::getType() const
     return interfaceType;
 }
 
-int vruiCollabInterface::composeMessage(const char messageType, const char *message)
+size_t vruiCollabInterface::composeMessage(const char messageType, const char *message)
 {
 
     size_t len;
@@ -134,10 +134,10 @@ int vruiCollabInterface::composeMessage(const char messageType, const char *mess
 void vruiCollabInterface::sendLockMessage(const char *message)
 {
 
-    int len = composeMessage('L', message);
+    size_t len = composeMessage('L', message);
 
     if (coim)
-        vruiRendererInterface::the()->sendCollabMessage(this, sendBuf, len);
+        vruiRendererInterface::the()->sendCollabMessage(this, sendBuf, (int)len);
 }
 
 /** send interactions to remote interfaces
