+++ src/module/general/ReadFamu/src/ResultsFileParserCovise.cpp
@@ -51,6 +51,6 @@ bool ResultsFileParserCovise::readTimeSteps(
             _results.push_back(newResults);
         }
     }
-    _noOfTimeSteps = _results.size();
+    _noOfTimeSteps = (int)_results.size();
     return true;
 }
